refactor and clean things up more
